+++ src/main/ts/badge/event/cheers.ts
@@ -1,4 +1,5 @@
-import { BadgeData } from 'coh-content-db'
+import { BadgeData, contactLink } from 'coh-content-db'
+import { Luna } from '../../contact/luna'
 
 export const Cheers: BadgeData = {
   type: 'event',
@@ -9,7 +10,7 @@ export const Cheers: BadgeData = {
   releaseDate: '2025-04-19',
   morality: 'all',
   badgeText: `You have helped celebrate the 21st anniversary of City of Heroes.`,
-  acquisition: `Awarded upon login during May 2025, and thereafter available for purchase from Luna in Ouroboros during the anniversary event in May.`,
+  acquisition: `Awarded upon login during May 2025, and thereafter available for purchase from ${contactLink(Luna)} during the anniversary event in May.`,
   links: [
     { title: 'Cheers!', href: 'https://homecoming.wiki/wiki/Cheers_Badge' },
   ],