+++ src/main/ts/badge/event/exuberant.ts
@@ -1,4 +1,5 @@
-import { BadgeData } from 'coh-content-db'
+import { BadgeData, contactLink } from 'coh-content-db'
+import { Luna } from '../../contact/luna'
 
 export const Exuberant: BadgeData = {
   type: 'event',
@@ -9,7 +10,7 @@ export const Exuberant: BadgeData = {
   releaseDate: '2026-04-27',
   morality: 'all',
   badgeText: `You have helped celebrate the 22nd anniversary of City of Heroes.`,
-  acquisition: `Awarded upon login during May 2026, and thereafter available for purchase from Luna in Ouroboros during the anniversary event in May.`,
+  acquisition: `Awarded upon login during May 2026, and thereafter available for purchase from ${contactLink(Luna)} during the anniversary event in May.`,
   links: [
     { title: 'Exuberant', href: 'https://homecoming.wiki/wiki/Exuberant_Badge' },
   ],