+++ src/main/ts/badge/event/lucky.ts
@@ -1,5 +1,5 @@
-import { BadgeData, zoneLink } from 'coh-content-db'
-import { Ouroboros } from '../../zone/ouroboros'
+import { BadgeData, contactLink } from 'coh-content-db'
+import { Luna } from '../../contact/luna'
 
 export const Lucky: BadgeData = {
   type: 'event',
@@ -10,7 +10,7 @@ export const Lucky: BadgeData = {
   releaseDate: '2012-11-30',
   morality: 'all',
   badgeText: 'You have helped celebrate the 7th anniversary of City of Heroes/Villains.',
-  acquisition: `Available for purchase from Luna in ${zoneLink(Ouroboros)} during the anniversary event in May.`,
+  acquisition: `Available for purchase from ${contactLink(Luna)} during the anniversary event in May.`,
   links: [
     { title: 'Lucky Badge', href: 'https://homecoming.wiki/wiki/Lucky_Badge' },
   ],