+++ src/main/ts/badge/event/resurgent.ts
@@ -1,4 +1,5 @@
-import { BadgeData } from 'coh-content-db'
+import { BadgeData, contactLink } from 'coh-content-db'
+import { Luna } from '../../contact/luna'
 
 export const Resurgent: BadgeData = {
   type: 'event',
@@ -9,7 +10,7 @@ export const Resurgent: BadgeData = {
   releaseDate: '2019-06-01',
   morality: 'all',
   badgeText: `You have helped celebrate the 16th anniversary of City of Heroes.`,
-  acquisition: `Available for purchase from Luna in Ouroboros during the anniversary event in May.`,
+  acquisition: `Available for purchase from ${contactLink(Luna)} during the anniversary event in May.`,
   links: [
     { title: 'Resurgent Badge', href: 'https://homecoming.wiki/wiki/Resurgent_Badge' },
   ],