+++ src/main/ts/badge/event/reunited.ts
@@ -1,5 +1,5 @@
-import { BadgeData, zoneLink } from 'coh-content-db'
-import { Ouroboros } from '../../zone/ouroboros'
+import { BadgeData, contactLink } from 'coh-content-db'
+import { Luna } from '../../contact/luna'
 
 export const Reunited: BadgeData = {
   type: 'event',
@@ -10,7 +10,7 @@ export const Reunited: BadgeData = {
   releaseDate: '2024-05-02',
   morality: 'all',
   badgeText: `You have helped celebrate the 20th anniversary of City of Heroes.`,
-  acquisition: `Awarded upon login during May 2024, and thereafter available for purchase from Luna in ${zoneLink(Ouroboros)} during the anniversary event in May.`,
+  acquisition: `Awarded upon login during May 2024, and thereafter available for purchase from ${contactLink(Luna)} during the anniversary event in May.`,
   links: [
     { title: 'Reunited Badge', href: 'https://homecoming.wiki/wiki/Reunited_Badge' },
   ],