+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.2.1] - 2026-04-01
+## [2.2.2] - 2026-04-28
 
 ### Fixed
 
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [2.2.2] Added Exuberant 2026 Anniversary Badge ([#196](https://github.com/n15g/coh-content-db-homecoming/pull/196))
+- [2.2.2] Added contact details for the anniversary vendor Luna.
 - [2.2.1] Included an April Fool's joke in the CHANGELOG
 - [2.2.0] `gameId` added to all badges to support new `/build_save` changes ([#192](https://github.com/n15g/coh-content-db-homecoming/pull/192))